requirements.txt
allure-pytest==2.13.4
allure-python-commons==2.13.4
attrs==22.2.0
Brlapi==0.8.4
certifi==2022.9.24
h11==0.14.0
idna==3.3
iniconfig==1.1.1
outcome==1.2.0
packaging==23.0
pluggy==1.0.0+repack
pycairo==1.20.1
pycups==2.0.1
pycurl==7.45.2
Pygments==2.14.0
PyGObject==3.42.2
PyJWT==2.6.0
pylibacl==0.7.0
pyparsing==3.0.9
pysmbc==1.0.23
pytest==7.2.1
python-apt==2.6.0
python-debian==0.1.49
python-debianbts==4.0.1
pyxattr==0.8.1
pyxdg==0.28
selenium==4.8.3
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
py~=1.11.0
cryptography~=38.0.4
pip~=23.0.1
distro~=1.8.0
wheel~=0.38.4
tornado~=6.2
xdg~=5
setuptools~=66.1.1
httplib2~=0.20.4
six~=1.16.0
oauthlib~=3.2.2
wadllib~=1.3.6
exceptiongroup~=1.1.0
coverage~=7.4.4
PyQt5~=5.15.9
chardet~=5.1.0
blinker~=1.5
Faker~=24.11.0